task-5: folders refactoring, change readme, package json
