+++ README.md
@@ -7,16 +7,21 @@ The `cdk.json` file tells the CDK Toolkit how to execute the app.
 ## Useful commands
 
 * `npx simple-git-hooks` activate hooks
-* `npm run deploy` deploy service
+* `npm run deploy` deploy services
+* `npm run deploy:import-service` deploy ImportServiceStack
+* `npm run deploy:product-service` deploy ProductServiceStack
 
 ## Another useful commands
 
 * `npm run build`   compile typescript to js
 * `npm run watch`   watch for changes and compile
 * `npm run test`    perform the jest unit tests
-* `cdk deploy`      deploy this stack to your default AWS account/region
+* `cdk deploy`      deploy all stack to your default AWS account/region
 * `cdk diff`        compare deployed stack with current state
 * `cdk synth`       emits the synthesized CloudFormation template
 
-## Linksg
-[Swagger](https://app.swaggerhub.com/apis/Ckpuna4/product-api/1)
+## Links
+* [Swagger](https://app.swaggerhub.com/apis/Ckpuna4/product-api/1)
+* [ImportServiceStack](https://5sa7tml0b4.execute-api.us-east-1.amazonaws.com/prod/)
+* [ProductServiceStack](https://m1rcin07mg.execute-api.us-east-1.amazonaws.com/prod/)
+* [CloudFront website](https://dgaojw28dgevx.cloudfront.net)
