+++ jest.config.js
@@ -1,6 +1,6 @@
 module.exports = {
   testEnvironment: 'node',
-  roots: ['<rootDir>/src/tests', '<rootDir>/src/import-service/tests'],
+  roots: ['<rootDir>/src/product-service/tests', '<rootDir>/src/import-service/tests'],
   testMatch: ['**/*.test.ts'],
   moduleNameMapper: {
     '^/opt/handlers-utils$': '<rootDir>/src/utils/handlers-utils',
