+++ package.json
@@ -9,7 +9,9 @@
     "watch": "tsc -w",
     "test": "jest",
     "cdk": "cdk",
-    "deploy": "npm run build && cdk deploy --profile skreepatch --outputs-file ./cdk-outputs.json",
+    "deploy": "npm run build && cdk deploy --all --profile skreepatch --outputs-file ./cdk-outputs.json",
+    "deploy:import-service": "npm run build && cdk deploy ImportServiceStack --profile skreepatch --outputs-file ./cdk-outputs.json",
+    "deploy:product-service": "npm run build && cdk deploy ProductServiceStack --profile skreepatch --outputs-file ./cdk-outputs.json",
     "init-tables": "tsc ./tools/add-items.ts && node ./tools/add-items.js"
   },
   "devDependencies": {
