+++ src/my-super-secret-backend-project.ts
@@ -1,11 +1,11 @@
 #!/usr/bin/env node
 import 'source-map-support/register';
 import * as cdk from 'aws-cdk-lib';
-import {ProductServiceStack} from './product-service-stack';
-import {ImportServiceStack} from "./import-service/import-service-stack";
+import {ProductServiceStack} from './product-service/product-service-stack';
+import {ImportServiceStack} from './import-service/import-service-stack';
 
 const app = new cdk.App();
 
-// new ProductServiceStack(app, 'ProductServiceStack');
+new ProductServiceStack(app, 'ProductServiceStack');
 new ImportServiceStack(app, 'ImportServiceStack');
