+++ src/product-service/handlers/create-product/index.ts
