+++ src/product-service/handlers/get-products-by-id/index.ts
