+++ src/product-service/handlers/get-products-list/index.ts
