+++ tsconfig.json
@@ -23,7 +23,6 @@
       "./node_modules/@types"
     ],
     "paths": {
-      "/opt/mock-data": ["./src/mocks/mock-data"],
       "/opt/handlers-utils": ["./src/utils/handlers-utils"]
     },
     "outDir": "./dist",
@@ -33,7 +32,7 @@
   "exclude": [
     "node_modules",
     "cdk.out",
-    "src/tests",
+    "src/product-service/tests",
     "src/import-service/tests",
     "tools"
   ]
